@@ -39,109 +39,110 @@
     "snyk-protect": "snyk protect"
   },
   "dependencies": {
-    "@angular-eslint/schematics": "^0.0.1-alpha.29",
-    "@angular/animations": "^9.0.2",
-    "@angular/common": "^9.0.2",
-    "@angular/compiler": "^9.0.2",
-    "@angular/core": "^9.0.2",
-    "@angular/forms": "^9.0.2",
-    "@angular/platform-browser": "^9.0.2",
-    "@angular/platform-browser-dynamic": "^9.0.2",
-    "@angular/router": "^9.0.2",
-    "@godaddy/terminus": "^4.3.1",
-    "@golevelup/nestjs-modules": "^0.4.0",
-    "@marcj/marshal": "^2.1.3",
-    "@marcj/marshal-nest": "^2.1.3",
-    "@nestjs/common": "^7.0.8",
-    "@nestjs/core": "^7.0.8",
+    "@angular-eslint/schematics": "^0.0.1-alpha.32",
+    "@angular/animations": "^9.1.11",
+    "@angular/common": "^9.1.11",
+    "@angular/compiler": "^9.1.11",
+    "@angular/core": "^9.1.11",
+    "@angular/forms": "^9.1.11",
+    "@angular/platform-browser": "^9.1.11",
+    "@angular/platform-browser-dynamic": "^9.1.11",
+    "@angular/router": "^9.1.11",
+    "@godaddy/terminus": "^4.4.1",
+    "@golevelup/nestjs-modules": "^0.4.1",
+    "@marcj/marshal": "^2.1.13",
+    "@marcj/marshal-nest": "^2.1.13",
+    "@nestjs/common": "^7.2.0",
+    "@nestjs/core": "^7.2.0",
     "@nestjs/passport": "^7.0.0",
-    "@nestjs/platform-express": "^7.0.8",
+    "@nestjs/platform-express": "^7.2.0",
     "@nestjs/terminus": "^7.0.1",
-    "@nrwl/angular": "^9.0.2",
+    "@nrwl/angular": "^9.4.5",
     "@ogma/logger": "^0.1.0",
     "@ogma/nestjs-module": "^0.1.0",
     "@ogma/platform-express": "^0.1.0",
-    "bcrypt": "^4.0.0",
+    "bcrypt": "^5.0.0",
     "compression": "^1.7.4",
     "connect-redis": "^4.0.4",
     "core-js": "^3.6.4",
     "dotenv": "^8.2.0",
-    "express-rate-limit": "^5.1.1",
+    "express-rate-limit": "^5.1.3",
     "express-session": "^1.17.0",
-    "helmet": "^3.21.2",
-    "nestjs-spelunker": "^0.1.0",
+    "helmet": "^3.23.2",
+    "nestjs-oauth": "^0.3.0",
+    "nestjs-spelunker": "^0.2.0",
     "passport": "^0.4.1",
     "passport-google-oauth": "^2.0.0",
     "passport-local": "^1.0.0",
-    "pg": "^8.0.3",
+    "pg": "^8.2.1",
     "redis": "^3.0.2",
     "reflect-metadata": "^0.1.12",
-    "runtypes": "^4.1.0",
+    "runtypes": "^4.3.0",
     "rxjs": "^6.5.4",
-    "snyk": "^1.296.0",
-    "sqlite3": "^4.1.1",
+    "snyk": "^1.348.2",
+    "sqlite3": "^4.2.0",
     "zone.js": "^0.10.2"
   },
   "devDependencies": {
     "@abraham/reflection": "^0.7.0",
-    "@angular-devkit/build-angular": "^0.901.3",
-    "@angular-eslint/builder": "0.0.1-alpha.29",
-    "@angular-eslint/eslint-plugin": "0.0.1-alpha.29",
-    "@angular-eslint/eslint-plugin-template": "0.0.1-alpha.29",
-    "@angular-eslint/template-parser": "0.0.1-alpha.29",
-    "@angular/cli": "^9.0.3",
-    "@angular/compiler-cli": "^9.0.2",
-    "@angular/language-service": "^9.0.2",
-    "@commitlint/cli": "^8.3.5",
-    "@commitlint/config-conventional": "^8.3.4",
+    "@angular-devkit/build-angular": "^0.901.9",
+    "@angular-eslint/builder": "0.0.1-alpha.32",
+    "@angular-eslint/eslint-plugin": "0.0.1-alpha.32",
+    "@angular-eslint/eslint-plugin-template": "0.0.1-alpha.32",
+    "@angular-eslint/template-parser": "0.0.1-alpha.32",
+    "@angular/cli": "^9.1.9",
+    "@angular/compiler-cli": "^9.1.11",
+    "@angular/language-service": "^9.1.11",
+    "@commitlint/cli": "^9.0.1",
+    "@commitlint/config-conventional": "^9.0.1",
     "@golevelup/nestjs-testing": "^0.1.2",
-    "@nestjs/schematics": "^7.0.0",
-    "@nestjs/testing": "^7.0.8",
-    "@nrwl/cypress": "^9.0.2",
-    "@nrwl/jest": "^9.0.2",
-    "@nrwl/nest": "^9.0.2",
-    "@nrwl/node": "^9.0.2",
-    "@nrwl/workspace": "^9.0.2",
+    "@nestjs/schematics": "^7.0.1",
+    "@nestjs/testing": "^7.2.0",
+    "@nrwl/cypress": "^9.4.5",
+    "@nrwl/jest": "^9.4.5",
+    "@nrwl/nest": "^9.4.5",
+    "@nrwl/node": "^9.4.5",
+    "@nrwl/workspace": "^9.4.5",
     "@types/bcrypt": "^3.0.0",
     "@types/compression": "^1.7.0",
-    "@types/connect-redis": "0.0.13",
+    "@types/connect-redis": "0.0.14",
     "@types/dotenv": "^8.2.0",
     "@types/express-rate-limit": "^5.0.0",
     "@types/express-session": "^1.15.16",
-    "@types/helmet": "0.0.45",
-    "@types/jest": "^25.1.3",
-    "@types/jquery": "^3.3.33",
-    "@types/node": "^13.7.4",
+    "@types/helmet": "0.0.47",
+    "@types/jest": "^26.0.3",
+    "@types/jquery": "^3.5.0",
+    "@types/node": "^14.0.14",
     "@types/passport": "^1.0.2",
     "@types/passport-google-oauth": "^1.0.41",
     "@types/passport-local": "^1.0.33",
     "@types/pg": "^7.14.1",
-    "@types/redis": "^2.8.16",
+    "@types/redis": "^2.8.22",
     "@types/sqlite3": "^3.1.6",
-    "@types/supertest": "^2.0.8",
-    "@typescript-eslint/eslint-plugin": "^2.30.0",
-    "@typescript-eslint/parser": "^2.30.0",
+    "@types/supertest": "^2.0.9",
+    "@typescript-eslint/eslint-plugin": "^3.4.0",
+    "@typescript-eslint/parser": "^3.4.0",
     "codelyzer": "^5.2.1",
-    "cypress": "^4.0.2",
-    "cz-conventional-changelog": "^3.1.0",
-    "eslint": "^6.8.0",
+    "cypress": "^4.9.0",
+    "cz-conventional-changelog": "^3.2.0",
+    "eslint": "^7.3.1",
     "eslint-config-prettier": "^6.11.0",
-    "eslint-plugin-cypress": "^2.10.3",
-    "eslint-plugin-prettier": "^3.1.3",
+    "eslint-plugin-cypress": "^2.11.1",
+    "eslint-plugin-prettier": "^3.1.4",
     "husky": "^4.2.3",
-    "jest": "^25.1.0",
-    "jest-preset-angular": "^8.0.0",
-    "lint-staged": "^10.0.7",
+    "jest": "^26.1.0",
+    "jest-preset-angular": "^8.2.1",
+    "lint-staged": "^10.2.11",
     "npm-check": "^5.9.0",
     "prettier": "^2.0.5",
     "supertest": "^4.0.2",
-    "ts-jest": "^25.2.1",
-    "ts-node": "^8.6.2",
-    "tslint": "^6.0.0",
+    "ts-jest": "^26.1.1",
+    "ts-node": "^8.10.2",
+    "tslint": "^6.1.2",
     "tslint-config-prettier": "^1.18.0",
     "tslint-sonarts": "^1.9.0",
-    "tslint-to-eslint-config": "^1.0.0-beta",
-    "typescript": "^3.8.2"
+    "tslint-to-eslint-config": "^1.0.0",
+    "typescript": "^3.9.5"
   },
   "engines": {
     "node": ">=10"
