@@ -6,12 +6,13 @@ import { DatabaseModule } from '../database/database.module';
 import { PassportModuleConfig } from '../options/passport.config';
 import { AuthController } from './auth/auth.controller';
 import { AuthService } from './auth/auth.service';
-import { GoogleUserService } from './google-user/google-user.service';
 import { GoogleStrategy } from './google.strategy';
 import { LocalStrategy } from './local.strategy';
 import { SessionSerializer } from './session.serializer';
 import { UserService } from './user/user.service';
 import { UsersController } from './user/user.controller';
+import { GoogleModule } from './google/google.module';
+import { OauthController } from './oauth.controller';
 
 @Module({
   imports: [
@@ -21,16 +22,16 @@ import { UsersController } from './user/user.controller';
     ConfigModule.Deferred,
     DatabaseModule.forFeature({ tableName: 'players' }),
     OgmaModule.forFeature(UserService),
+    GoogleModule,
   ],
   providers: [
     AuthService,
+    GoogleStrategy,
     LocalStrategy,
     SessionSerializer,
-    GoogleStrategy,
-    GoogleUserService,
     UserService,
   ],
   exports: [PassportModule, AuthService],
-  controllers: [AuthController, UsersController],
+  controllers: [AuthController, OauthController, UsersController],
 })
 export class AuthModule {}
