@@ -1,6 +1,8 @@
 import { NestFactory } from '@nestjs/core';
 import { OgmaService } from '@ogma/nestjs-module';
 import { SpelunkerModule } from 'nestjs-spelunker';
+import { writeFileSync } from 'fs';
+import { join } from 'path';
 
 import { AppModule } from './app/app.module';
 import { ConfigService } from './app/config/config.service';
@@ -15,7 +17,10 @@ async function bootstrap() {
   app.useLogger(logger);
   const port = config.port;
   configure(app, config, logger);
-  SpelunkerModule.explore(app, logger);
+  writeFileSync(
+    join(process.cwd(), 'server.json'),
+    Buffer.from(JSON.stringify(SpelunkerModule.explore(app))),
+  );
   await app.listen(port);
   logger.log(`Listening at ${await app.getUrl()}`, 'NestApplication');
 }
